requirements-uv.txt
# Runtime dependencies for uv-based installs (PyTorch is installed separately
# so you can pick the right build for your hardware).
absl-py==0.14.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
audioread==3.0.0
av==8.0.3
black==21.4b2
cachetools==4.2.2
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
cython==0.29.24
einops==0.6.0
fairscale==0.4.1
filelock==3.8.0
future==0.18.2
fvcore==0.1.5.post20211023
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
huggingface-hub==0.9.1
hydra-core==1.1.1
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.12.0
importlib-resources==5.2.2
iopath==0.1.8
joblib==1.1.0
kiwisolver==1.3.2
librosa==0.9.2
llvmlite==0.39.0
markdown==3.3.4
matplotlib==3.4.3
moviepy==1.0.3
mypy-extensions==0.4.3
networkx==2.6.3
numba==0.56.0
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.3.56
pandas==1.3.3
parameterized==0.8.1
pathspec==0.9.0
pillow==8.3.2
pooch==1.6.0
portalocker==2.3.2
protobuf==3.18.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytorchvideo==0.1.5
pytz==2021.1
pyyaml==5.4.1
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.4.0
rsa==4.7.2
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
simplejson==3.17.5
six==1.16.0
soundfile==0.10.3.post1
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
threadpoolctl==3.1.0
timm==0.6.12
tokenizers==0.12.1
toml==0.10.2
tqdm==4.62.3
transformers==4.21.3
typing_extensions==3.10.0.2
Werkzeug==2.0.1
yacs==0.1.8
zipp==3.5.0
